Take ACE_Settings out of config.cpp
