+++ addons/explosives/ACE_Settings.hpp
@@ -1,0 +1,10 @@
+class ACE_Settings {
+    class GVAR(RequireSpecialist) {
+        value = 0;
+        typeName = "BOOL";
+    };
+    class GVAR(PunishNonSpecialists) {
+        value = 1;
+        typeName = "BOOL";
+    };
+};
