+++ addons/goggles/ACE_Settings.hpp
@@ -1,0 +1,8 @@
+class ACE_Settings {
+    class GVAR(showInThirdPerson) {
+        value = 0;
+        typeName = "BOOL";
+        isClientSettable = 1;
+        displayName = "$STR_ACE_Goggles_ShowInThirdPerson";
+    };
+};
