+++ addons/goggles/config.cpp
@@ -243,14 +243,7 @@ class SniperCloud {
     ACE_Goggles_BulletCount = 1;
 };
 
-class ACE_Settings {
-    class GVAR(showInThirdPerson) {
-        value = 0;
-        typeName = "BOOL";
-        isClientSettable = 1;
-        displayName = "$STR_ACE_Goggles_ShowInThirdPerson";
-    };
-};
+#include "ACE_Settings.hpp"
 
 class CfgCloudlets {
     class Default;
