+++ addons/interaction/ACE_Settings.hpp
@@ -1,0 +1,6 @@
+class ACE_Settings {
+  class GVAR(EnableTeamManagement) {
+    value = 1;
+    typeName = "BOOL";
+  };
+};
