+++ addons/inventory/config.cpp
@@ -13,15 +13,4 @@ class CfgPatches {
 };
 
 #include "RscDisplayInventory.hpp"
-
-class ACE_Settings {
-    class GVAR(inventoryDisplaySize) {
-        value = 0;
-        typeName = "SCALAR";
-        isClientSettable = 1;
-        displayName = "$STR_ACE_Inventory_SettingName";
-        description = "$STR_ACE_Inventory_SettingDescription";
-        values[] = {"$str_medium", "$str_large", "$str_very_large"};
-    };
-};
-
+#include "ACE_Settings.hpp"
