+++ addons/map/ACE_Settings.hpp
@@ -1,0 +1,30 @@
+class ACE_Settings {
+    class GVAR(BFT_Interval) {
+        value = 1.0;
+        typeName = "SCALAR";
+    };
+    class GVAR(BFT_Enabled) {
+        value = 0;
+        typeName = "BOOL";
+    };
+    class GVAR(BFT_HideAiGroups) {
+        value = 0;
+        typeName = "BOOL";
+    };
+    class GVAR(mapIllumination) {
+        value = 1;
+        typeName = "BOOL";
+    };
+    class GVAR(mapShake) {
+        value = 1;
+        typeName = "BOOL";
+    };
+    class GVAR(mapLimitZoom) {
+        value = 0;
+        typeName = "BOOL";
+    };
+    class GVAR(mapShowCursorCoordinates) {
+        value = 0;
+        typeName = "BOOL";
+    };
+};
