+++ addons/map/config.cpp
@@ -23,37 +23,7 @@ class RscButtonMenuCancel;
 class RscButtonMenu;
 class RscEdit;
 
-class ACE_Settings {
-    class GVAR(BFT_Interval) {
-        value = 1.0;
-        typeName = "SCALAR";
-    };
-    class GVAR(BFT_Enabled) {
-        value = 0;
-        typeName = "BOOL";
-    };
-    class GVAR(BFT_HideAiGroups) {
-        value = 0;
-        typeName = "BOOL";
-    };
-    class GVAR(mapIllumination) {
-        value = 1;
-        typeName = "BOOL";
-    };
-    class GVAR(mapShake) {
-        value = 1;
-        typeName = "BOOL";
-    };
-    class GVAR(mapLimitZoom) {
-        value = 0;
-        typeName = "BOOL";
-    };
-    class GVAR(mapShowCursorCoordinates) {
-        value = 0;
-        typeName = "BOOL";
-    };
-};
-
+#include "ACE_Settings.hpp"
 #include "CfgEventHandlers.hpp"
 #include "CfgMarkers.hpp"
 #include "CfgVehicles.hpp"
