+++ addons/microdagr/ACE_Settings.hpp
@@ -1,0 +1,7 @@
+class ACE_Settings {
+    class GVAR(MapDataAvailable) {
+        value = 2;
+        typeName = "SCALAR";
+        isClientSettable = 0;
+    };
+};
