+++ addons/optionsmenu/config.cpp
@@ -27,16 +27,7 @@ class CfgAddons {
 #include "gui\pauseMenu.hpp"
 
 #include "CfgVehicles.hpp"
-
-class ACE_Settings {
-    class GVAR(optionMenuDisplaySize) {
-        value = 0;
-        typeName = "SCALAR";
-        isClientSettable = 1;
-        displayName = "$STR_ACE_optionsMenu_uiScaing";
-        values[] = {"$str_medium", "$str_large", "$str_very_large"};
-    };
-};
+#include "ACE_Settings.hpp"
 
 class ACE_Extensions {
     extensions[] += {"ace_clipboard"};
