+++ addons/reload/ACE_Settings.hpp
@@ -1,0 +1,9 @@
+class ACE_Settings {
+    class GVAR(DisplayText) {
+        typeName = "BOOL";
+        isClientSettable = 1;
+        value = 1;
+        displayName = "$STR_ACE_reload_SettingDisplayTextName";
+        description = "$STR_ACE_reload_SettingDisplayTextDesc";
+    };
+};
