+++ addons/vehiclelock/ACE_Settings.hpp
@@ -1,0 +1,14 @@
+class ACE_Settings {
+    class GVAR(DefaultLockpickStrength) {
+        value = 10;
+        typeName = "SCALAR";
+    };
+    class GVAR(LockVehicleInventory) {
+        value = 0;
+        typeName = "BOOL";
+    };    
+    class GVAR(VehicleStartingLockState) {
+        value = -1;
+        typeName = "SCALAR";
+    };
+};
